Added resource about CSS frameworks
I added a link to a roundup where 41 developers give their opinion about the best CSS frameworks. The roundup also explains in great detail what css frameworks (and grid systems) are. There also is a top 5 best CSS frameworks, based on the insights of the developers.

I think the roundup is a great resource because it goes deeper into CSS frameworks.

I hope this helps.